fix files location
