--- src/stores/LLMEnums.py
@@ -1,11 +1,0 @@
-from enum import Enum
-
-class LLMEnums(Enum):
-    OPENAI = "OPENAI"
-    COHERE = "COHERE"
-
-class OpenAIEnums(Enum):
-    SYSTEM = "system"
-    USER = "user"
-    ASSISTANT = "assistant"
-
